tests/wpt/web-platform-tests/css/css-toggle/parsing/toggle-group-valid.tentative.html

<!DOCTYPE HTML>
<meta charset="UTF-8">
<title>CSS Toggles: parsing toggle-group with valid values</title>
<link rel="author" title="L. David Baron" href="https://dbaron.org/">
<link rel="author" title="Google" href="http://www.google.com/">
<link rel="help" href="https://tabatkins.github.io/css-toggle/#toggle-group-property">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>

<body>
<script>

test_valid_value('toggle-group', 'none');
test_valid_value('toggle-group', 'self self');
test_valid_value('toggle-group', 'mytoggle');
test_valid_value('toggle-group', 'mytoggle, yourtoggle');
test_valid_value('toggle-group', 'mytoggle, mytoggle');
test_valid_value('toggle-group', 'mytoggle self, yourtoggle');
test_valid_value('toggle-group', 'mytoggle self , yourtoggle self', 'mytoggle self, yourtoggle self');
test_valid_value('toggle-group', 'mytoggle,yourtoggle self', 'mytoggle, yourtoggle self');

</script>
